+++ plugins/catalog-backend/src/database/operations/provider/deleteWithEagerPruningOfChildren.test.ts
@@ -18,7 +18,11 @@ import { TestDatabaseId, TestDatabases } from '@backstage/backend-test-utils';
 import { Knex } from 'knex';
 import * as uuid from 'uuid';
 import { applyDatabaseMigrations } from '../../migrations';
-import { DbRefreshStateReferencesRow, DbRefreshStateRow } from '../../tables';
+import {
+  DbRefreshStateReferencesRow,
+  DbRefreshStateRow,
+  DbRelationsRow,
+} from '../../tables';
 import { deleteWithEagerPruningOfChildren } from './deleteWithEagerPruningOfChildren';
 
 jest.setTimeout(60_000);
@@ -43,6 +47,22 @@ describe('deleteWithEagerPruningOfChildren', () => {
     );
   }
 
+  async function insertRelation(
+    knex: Knex,
+    ...relations: { from: string; to: string }[]
+  ) {
+    for (const rel of relations) {
+      await knex<DbRelationsRow>('relations').insert({
+        originating_entity_id: await knex<DbRefreshStateRow>('refresh_state')
+          .select('entity_id')
+          .then(rows => rows[0].entity_id), // doesn't matter which one, this is consumed pre-deletion
+        source_entity_ref: rel.from,
+        target_entity_ref: rel.to,
+        type: 'fake',
+      });
+    }
+  }
+
   async function insertEntity(knex: Knex, ...entityRefs: string[]) {
     for (const ref of entityRefs) {
       await knex<DbRefreshStateRow>('refresh_state').insert({
@@ -64,6 +84,14 @@ describe('deleteWithEagerPruningOfChildren', () => {
     return rows.map(r => r.entity_ref);
   }
 
+  async function entitiesMarkedForStitching(knex: Knex) {
+    const rows = await knex<DbRefreshStateRow>('refresh_state')
+      .orderBy('entity_ref')
+      .select('entity_ref')
+      .where('result_hash', '=', 'force-stitching');
+    return rows.map(r => r.entity_ref);
+  }
+
   it.each(databases.eachSupportedId())(
     'works for the simple path, %p',
     async databaseId => {
@@ -78,7 +106,7 @@ describe('deleteWithEagerPruningOfChildren', () => {
 
           Scenario: P1 issues delete for E1 and E3
 
-          Result: E1, E2, and E3 deleted; E4 and E5 remain
+          Result: E1, E2, and E3 deleted; E4 and E5 remain; E4 marked for stitching because it had a relation to a deleted entity
        */
       const knex = await createDatabase(databaseId);
       await insertEntity(knex, 'E1', 'E2', 'E3', 'E4', 'E5');
@@ -90,12 +118,14 @@ describe('deleteWithEagerPruningOfChildren', () => {
         { source_key: 'P1', target_entity_ref: 'E4' },
         { source_key: 'P2', target_entity_ref: 'E5' },
       );
+      await insertRelation(knex, { from: 'E4', to: 'E2' });
       await deleteWithEagerPruningOfChildren({
         knex,
         sourceKey: 'P1',
         entityRefs: ['E1', 'E3'],
       });
       await expect(remainingEntities(knex)).resolves.toEqual(['E4', 'E5']);
+      await expect(entitiesMarkedForStitching(knex)).resolves.toEqual(['E4']);
     },
   );
 
@@ -113,7 +143,7 @@ describe('deleteWithEagerPruningOfChildren', () => {
 
           Scenario: P1 issues delete for E1
 
-          Result: E1 deleted; E2 remains
+          Result: E1 deleted; E2 remains; E2 marked for stitching because it had a relation to a deleted entity
        */
       const knex = await createDatabase(databaseId);
       await insertEntity(knex, 'E1', 'E2');
@@ -123,12 +153,14 @@ describe('deleteWithEagerPruningOfChildren', () => {
         { source_key: 'P1', target_entity_ref: 'E1' },
         { source_key: 'P1', target_entity_ref: 'E2' },
       );
+      await insertRelation(knex, { from: 'E2', to: 'E1' });
       await deleteWithEagerPruningOfChildren({
         knex,
         sourceKey: 'P1',
         entityRefs: ['E1'],
       });
       await expect(remainingEntities(knex)).resolves.toEqual(['E2']);
+      await expect(entitiesMarkedForStitching(knex)).resolves.toEqual(['E2']);
     },
   );
 
@@ -144,7 +176,7 @@ describe('deleteWithEagerPruningOfChildren', () => {
 
           Scenario: P1 issues delete for E1
 
-          Result: E1 deleted; E2 and E3 remain
+          Result: E1 deleted; E2 and E3 remain; E2 marked for stitching because it had a relation to a deleted entity
        */
       const knex = await createDatabase(databaseId);
       await insertEntity(knex, 'E1', 'E2', 'E3');
@@ -155,12 +187,18 @@ describe('deleteWithEagerPruningOfChildren', () => {
         { source_key: 'P2', target_entity_ref: 'E3' },
         { source_entity_ref: 'E3', target_entity_ref: 'E2' },
       );
+      await insertRelation(
+        knex,
+        { from: 'E2', to: 'E1' },
+        { from: 'E3', to: 'E2' },
+      );
       await deleteWithEagerPruningOfChildren({
         knex,
         sourceKey: 'P1',
         entityRefs: ['E1'],
       });
       await expect(remainingEntities(knex)).resolves.toEqual(['E2', 'E3']);
+      await expect(entitiesMarkedForStitching(knex)).resolves.toEqual(['E2']);
     },
   );
 
@@ -193,6 +231,7 @@ describe('deleteWithEagerPruningOfChildren', () => {
         entityRefs: ['E1'],
       });
       await expect(remainingEntities(knex)).resolves.toEqual(['E2', 'E3']);
+      await expect(entitiesMarkedForStitching(knex)).resolves.toEqual([]);
     },
   );
 
@@ -208,7 +247,7 @@ describe('deleteWithEagerPruningOfChildren', () => {
 
           Scenario: P1 issues delete for E1, then for E3
 
-          Result: Everything deleted, but in two steps
+          Result: Everything deleted, but in two steps; E4 marked for stitching in the first step because it had a relation to a deleted entity
        */
       const knex = await createDatabase(databaseId);
       await insertEntity(knex, 'E1', 'E2', 'E3', 'E4', 'E5', 'E6');
@@ -224,6 +263,7 @@ describe('deleteWithEagerPruningOfChildren', () => {
         { source_entity_ref: 'E3', target_entity_ref: 'E5' },
         { source_entity_ref: 'E5', target_entity_ref: 'E6' },
       );
+      await insertRelation(knex, { from: 'E4', to: 'E2' });
       await deleteWithEagerPruningOfChildren({
         knex,
         sourceKey: 'P1',
@@ -235,12 +275,14 @@ describe('deleteWithEagerPruningOfChildren', () => {
         'E5',
         'E6',
       ]);
+      await expect(entitiesMarkedForStitching(knex)).resolves.toEqual(['E4']);
       await deleteWithEagerPruningOfChildren({
         knex,
         sourceKey: 'P1',
         entityRefs: ['E3'],
       });
       await expect(remainingEntities(knex)).resolves.toEqual([]);
+      await expect(entitiesMarkedForStitching(knex)).resolves.toEqual([]);
     },
   );
 
@@ -277,6 +319,7 @@ describe('deleteWithEagerPruningOfChildren', () => {
         'E2',
         'E4',
       ]);
+      await expect(entitiesMarkedForStitching(knex)).resolves.toEqual([]);
     },
   );
 });