+++ packages/backend-plugin-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/backend-plugin-manager
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-backend@1.13.1
+
 ## 0.0.1
 
 ### Patch Changes