+++ plugins/catalog-backend-module-github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-catalog-backend-module-github
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-backend@1.13.1
+
 ## 0.4.0
 
 ### Minor Changes